+++ opyoid/providers/providers_factories/list_provider_factory.py
@@ -16,6 +16,6 @@ class ListProviderFactory(ProviderFactory):
         return TypeChecker.is_list(context.target.type)
 
     def create(self, context: InjectionContext[List[InjectedT]]) -> Provider[List[InjectedT]]:
-        new_target = Target(context.target.type.__args__[0], context.target.annotation)
+        new_target = Target(context.target.type.__args__[0], context.target.named)
         new_context = context.get_child_context(new_target)
         return ListProvider([new_context.get_provider()])